:art: Style: activity 파일 이동
<footer>
- 관련: #13
