+++ app/src/main/java/com/example/saybettereducator/ui/activity/MainActivity.kt
@@ -1,4 +1,4 @@
-package com.example.saybettereducator.ui
+package com.example.saybettereducator.ui.activity
 
 import android.Manifest
 import android.content.Intent
@@ -11,29 +11,21 @@ import androidx.activity.compose.setContent
 import androidx.activity.result.contract.ActivityResultContracts
 import androidx.annotation.RequiresApi
 import androidx.compose.foundation.Image
-import androidx.compose.foundation.background
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Box
-import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.Spacer
-import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.layout.width
-import androidx.compose.foundation.rememberScrollState
-import androidx.compose.foundation.verticalScroll
 import androidx.compose.material3.ExperimentalMaterial3Api
 import androidx.compose.material3.Icon
-import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.NavigationBar
 import androidx.compose.material3.NavigationBarItem
 import androidx.compose.material3.Scaffold
-import androidx.compose.material3.Surface
 import androidx.compose.material3.Text
 import androidx.compose.material3.TopAppBar
-import androidx.compose.material3.TopAppBarColors
 import androidx.compose.material3.TopAppBarDefaults
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.MutableState
@@ -42,12 +34,9 @@ import androidx.compose.runtime.mutableStateOf
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.Color
-import androidx.compose.ui.layout.onGloballyPositioned
-import androidx.compose.ui.layout.positionInRoot
 import androidx.compose.ui.platform.LocalContext
 import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.text.font.FontFamily
-import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
@@ -56,17 +45,15 @@ import androidx.navigation.compose.composable
 import androidx.navigation.compose.currentBackStackEntryAsState
 import androidx.navigation.compose.rememberNavController
 import com.example.saybettereducator.R
-import com.example.saybettereducator.model.remote.dto.DataModel
-import com.example.saybettereducator.ui.calendar.CalendarScreen
-import com.example.saybettereducator.ui.home.HomeScreen
-import com.example.saybettereducator.ui.learner.LearnerScreen
-import com.example.saybettereducator.ui.solution.SolutionScreen
+import com.example.saybettereducator.data.model.DataModel
+import com.example.saybettereducator.ui.view.screen.CalendarScreen
+import com.example.saybettereducator.ui.view.screen.HomeScreen
+import com.example.saybettereducator.ui.view.screen.LearnerScreen
+import com.example.saybettereducator.ui.view.screen.SolutionScreen
 import com.example.saybettereducator.ui.theme.GrayW40
 import com.example.saybettereducator.ui.theme.MainGreen
-import com.example.saybettereducator.ui.theme.SaybetterEducatorTheme
 import com.example.saybettereducator.ui.theme.montserratFont
 import com.example.saybettereducator.ui.theme.pretendardMediumFont
-import com.example.saybettereducator.ui.videoCall.VideoCallActivity
 import com.example.saybettereducator.utils.permission.checkAndRequestPermissions
 import com.example.saybettereducator.utils.webrtc.repository.MainRepository
 import com.example.saybettereducator.utils.webrtc.service.MainService
@@ -78,20 +65,22 @@ import javax.inject.Inject
 class MainActivity : ComponentActivity(), MainService.CallEventListener {
 
     data class TestDialogState(
-        val isClick : Boolean = false,
+        val isClick: Boolean = false,
         val onClickSure: () -> Unit = {},
         val onClickCancel: () -> Unit = {},
     )
 
     private val customAlertDialogState = mutableStateOf(TestDialogState())
 
-    private var userid : String? = null
+    private var userid: String? = null
     private var currentReceivedModel: DataModel? = null
     private val testUser: String = "helloYI"
 
     //Hilt 의존성 주입
-    @Inject lateinit var mainRepository : MainRepository
-    @Inject lateinit var mainServiceRepository : MainServiceRepository
+    @Inject
+    lateinit var mainRepository: MainRepository
+    @Inject
+    lateinit var mainServiceRepository: MainServiceRepository
 
     @RequiresApi(Build.VERSION_CODES.O)
     override fun onCreate(savedInstanceState: Bundle?) {
@@ -103,9 +92,9 @@ class MainActivity : ComponentActivity(), MainService.CallEventListener {
     }
 
     @RequiresApi(Build.VERSION_CODES.O)
-    private fun init(){
+    private fun init() {
         userid = intent.getStringExtra("userid")
-        if(userid == null) finish()
+        if (userid == null) finish()
 
         MainService.listener = this
         startMyService()
@@ -121,7 +110,6 @@ class MainActivity : ComponentActivity(), MainService.CallEventListener {
     fun MainScreenPreview() {
         MainScreen()
     }
-
 
 
     @OptIn(ExperimentalMaterial3Api::class)
@@ -163,14 +151,16 @@ class MainActivity : ComponentActivity(), MainService.CallEventListener {
             topBar = {
                 Box {
                     TopAppBar(
-                        title = { Text(
-                            text = "Say Better",
-                            fontSize = 16.sp,
-                            fontFamily = FontFamily(montserratFont),
-                            color = MainGreen,
-                            modifier = Modifier
-                                .padding(start = 34.04.dp, top = 15.29.dp)
-                        ) },
+                        title = {
+                            Text(
+                                text = "Say Better",
+                                fontSize = 16.sp,
+                                fontFamily = FontFamily(montserratFont),
+                                color = MainGreen,
+                                modifier = Modifier
+                                    .padding(start = 34.04.dp, top = 15.29.dp)
+                            )
+                        },
                         colors = TopAppBarDefaults.topAppBarColors(
                             containerColor = Color.White
                         ),
@@ -204,14 +194,14 @@ class MainActivity : ComponentActivity(), MainService.CallEventListener {
                             Icon(
                                 painter = painterResource(id = R.drawable.home_nav),
                                 contentDescription = "bottom navigation home",
-                                tint = if(currentRoute=="home") MainGreen else GrayW40
+                                tint = if (currentRoute == "home") MainGreen else GrayW40
                             )
                         },
                         label = {
                             Text(
                                 text = "홈",
                                 fontFamily = FontFamily(pretendardMediumFont),
-                                color = if(currentRoute=="home") MainGreen else GrayW40,
+                                color = if (currentRoute == "home") MainGreen else GrayW40,
                                 fontSize = 12.sp,
                             )
                         }
@@ -223,14 +213,14 @@ class MainActivity : ComponentActivity(), MainService.CallEventListener {
                             Icon(
                                 painter = painterResource(id = R.drawable.learner_nav),
                                 contentDescription = "bottom navigation learner",
-                                tint = if(currentRoute=="learner") MainGreen else GrayW40
+                                tint = if (currentRoute == "learner") MainGreen else GrayW40
                             )
                         },
                         label = {
                             Text(
                                 text = "학습자",
                                 fontFamily = FontFamily(pretendardMediumFont),
-                                color = if(currentRoute=="learner") MainGreen else GrayW40,
+                                color = if (currentRoute == "learner") MainGreen else GrayW40,
                                 fontSize = 12.sp
                             )
                         }
@@ -242,14 +232,14 @@ class MainActivity : ComponentActivity(), MainService.CallEventListener {
                             Icon(
                                 painter = painterResource(id = R.drawable.solution_nav),
                                 contentDescription = "bottom navigation solution",
-                                tint = if(currentRoute=="solution") MainGreen else GrayW40
+                                tint = if (currentRoute == "solution") MainGreen else GrayW40
                             )
                         },
                         label = {
                             Text(
                                 text = "솔루션",
                                 fontFamily = FontFamily(pretendardMediumFont),
-                                color = if(currentRoute=="solution") MainGreen else GrayW40,
+                                color = if (currentRoute == "solution") MainGreen else GrayW40,
                                 fontSize = 12.sp
                             )
                         }
@@ -261,14 +251,14 @@ class MainActivity : ComponentActivity(), MainService.CallEventListener {
                             Icon(
                                 painter = painterResource(id = R.drawable.calendar_nav),
                                 contentDescription = "bottom navigation calendar",
-                                tint = if(currentRoute=="calendar") MainGreen else GrayW40
+                                tint = if (currentRoute == "calendar") MainGreen else GrayW40
                             )
                         },
                         label = {
                             Text(
                                 text = "캘린더",
                                 fontFamily = FontFamily(pretendardMediumFont),
-                                color = if(currentRoute=="calendar") MainGreen else GrayW40,
+                                color = if (currentRoute == "calendar") MainGreen else GrayW40,
                                 fontSize = 12.sp
                             )
                         }
@@ -306,9 +296,9 @@ class MainActivity : ComponentActivity(), MainService.CallEventListener {
     }
 
     //Video call 클릭되었을 때
-    private fun StartVideoCall(targetUserid : String, isCaller: Boolean) {
+    private fun StartVideoCall(targetUserid: String, isCaller: Boolean) {
         mainRepository.sendConnectionRequest(targetUserid) {
-            if(it) {
+            if (it) {
                 //videocall 시작해야함
                 //educator 되면 수정
                 intent = Intent(this@MainActivity, VideoCallActivity::class.java)
